Silenced tar warnings

+++ Compression/NSFileManager+Tar.m
@@ -231,7 +231,7 @@ static NSString * const kNSFileManagerLightUntarCorruptFileMessage = @"Invalid b
 {
     char type;
 
-    memcpy(&type, [self dataForObject:object inRange:NSMakeRange(offset + TAR_TYPE_POSITION, 1) orLocation:offset + TAR_TYPE_POSITION andLength:1].bytes, 1);
+    memcpy(&type, [self dataForObject:object inRange:NSMakeRange((unsigned int)offset + TAR_TYPE_POSITION, 1) orLocation:offset + TAR_TYPE_POSITION andLength:1].bytes, 1);
     return type;
 }
 
@@ -240,7 +240,7 @@ static NSString * const kNSFileManagerLightUntarCorruptFileMessage = @"Invalid b
     char nameBytes[TAR_NAME_SIZE + 1]; // TAR_NAME_SIZE+1 for nul char at end
 
     memset(&nameBytes, '\0', TAR_NAME_SIZE + 1); // Fill byte array with nul char
-    memcpy(&nameBytes, [self dataForObject:object inRange:NSMakeRange(offset + TAR_NAME_POSITION, TAR_NAME_SIZE) orLocation:offset + TAR_NAME_POSITION andLength:TAR_NAME_SIZE].bytes, TAR_NAME_SIZE);
+    memcpy(&nameBytes, [self dataForObject:object inRange:NSMakeRange((unsigned int)offset + TAR_NAME_POSITION, TAR_NAME_SIZE) orLocation:offset + TAR_NAME_POSITION andLength:TAR_NAME_SIZE].bytes, TAR_NAME_SIZE);
     return [NSString stringWithCString:nameBytes encoding:NSASCIIStringEncoding];
 }
 
@@ -249,14 +249,14 @@ static NSString * const kNSFileManagerLightUntarCorruptFileMessage = @"Invalid b
     char sizeBytes[TAR_SIZE_SIZE + 1]; // TAR_SIZE_SIZE+1 for nul char at end
 
     memset(&sizeBytes, '\0', TAR_SIZE_SIZE + 1); // Fill byte array with nul char
-    memcpy(&sizeBytes, [self dataForObject:object inRange:NSMakeRange(offset + TAR_SIZE_POSITION, TAR_SIZE_SIZE) orLocation:offset + TAR_SIZE_POSITION andLength:TAR_SIZE_SIZE].bytes, TAR_SIZE_SIZE);
+    memcpy(&sizeBytes, [self dataForObject:object inRange:NSMakeRange((unsigned int)offset + TAR_SIZE_POSITION, TAR_SIZE_SIZE) orLocation:offset + TAR_SIZE_POSITION andLength:TAR_SIZE_SIZE].bytes, TAR_SIZE_SIZE);
     return strtol(sizeBytes, NULL, 8); // Size is an octal number, convert to decimal
 }
 
 - (void)writeFileDataForObject:(id)object atLocation:(unsigned long long)location withLength:(unsigned long long)length atPath:(NSString *)path
 {
     if ([object isKindOfClass:[NSData class]]) {
-        [self createFileAtPath:path contents:[object subdataWithRange:NSMakeRange(location, length)] attributes:nil]; //Write the file on filesystem
+        [self createFileAtPath:path contents:[object subdataWithRange:NSMakeRange((unsigned int)location, (unsigned int)length)] attributes:nil]; //Write the file on filesystem
     } else if ([object isKindOfClass:[NSFileHandle class]]) {
         if ([[NSData data] writeToFile:path atomically:NO]) {
             NSFileHandle *destinationFile = [NSFileHandle fileHandleForWritingAtPath:path];
@@ -266,12 +266,12 @@ static NSString * const kNSFileManagerLightUntarCorruptFileMessage = @"Invalid b
 
             while (length > maxSize) {
                 @autoreleasepool {
-                    [destinationFile writeData:[object readDataOfLength:maxSize]];
+                    [destinationFile writeData:[object readDataOfLength:(unsigned int)maxSize]];
                     location += maxSize;
                     length -= maxSize;
                 }
             }
-            [destinationFile writeData:[object readDataOfLength:length]];
+            [destinationFile writeData:[object readDataOfLength:(unsigned int)length]];
             [destinationFile closeFile];
         }
     }
@@ -283,7 +283,7 @@ static NSString * const kNSFileManagerLightUntarCorruptFileMessage = @"Invalid b
         return [object subdataWithRange:range];
     } else if ([object isKindOfClass:[NSFileHandle class]]) {
         [object seekToFileOffset:location];
-        return [object readDataOfLength:length];
+        return [object readDataOfLength:(unsigned int)length];
     }
 
     return nil;
